requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7
pydantic==2.5.3
numpy==1.26.3
scikit-learn==1.4.0
python-docx==1.1.0
PyPDF2==3.0.1
langchain==0.1.0
openai==1.13.0
python-dotenv==1.0.0
